Rename pre06 package
