+++ cmd/lifecycle/platform/v03through05/platform.go
@@ -1,4 +1,4 @@
-package pre06
+package v03through05
 
 import (
 	"github.com/buildpacks/lifecycle/api"
